IN-196 feat: add styles to minis
