+++ src/features/createPost/ui/cropping/PhotoCrop.tsx
@@ -7,11 +7,10 @@ import { ControllersPanel } from '@/features/createPost/ui/cropping/ControllersP
 import { observer } from 'mobx-react-lite'
 
 type Props = {
-  goToSlide: (index: number) => void
   photo: PostPhoto
 }
 
-export const PhotoCrop = observer(({ goToSlide, photo }: Props) => {
+export const PhotoCrop = observer(({ photo }: Props) => {
   const addZoom = addPostStore.addZoom
   const addCrop = addPostStore.addCrop
 
@@ -38,10 +37,7 @@ export const PhotoCrop = observer(({ goToSlide, photo }: Props) => {
         onZoomChange={onZoom}
         zoom={photo.zoom ?? 1}
       />
-      <ControllersPanel
-        goToSlide={goToSlide}
-        id={photo.id}
-      />
+      <ControllersPanel id={photo.id} />
     </>
   )
 })
